advanced-api-project/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter==24.3

